requirements.txt
tensorflow==2.15.0
numpy
Pillow
gradio>=4.0
pandas
pyarrow
